refactor: group unit tests into "__tests__" directory
